Move other role related functions to roles.py
