+++ cogs/info.py
@@ -22,54 +22,10 @@ import discord
 import subprocess
 from discord.ext import commands
 
-from .utils.paginator import Pages
-
 
 class Info(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
-
-    @commands.command()
-    async def roles(self, ctx):
-        """Returns list of all roles in server"""
-        roleNames = list(map(lambda role: role.name + "\n", ctx.guild.roles))
-        roleNames.reverse()
-        p = Pages(ctx,
-                  item_list=roleNames,
-                  title="All roles in server",
-                  display_option=(3, 20),
-                  editable_content=False)
-
-        await p.paginate()
-
-    @commands.command()
-    async def inrole(self, ctx, *, queryRole):
-        """Returns list of users in the specified role"""
-        members = None
-        for role in ctx.guild.roles:
-            if role.name.lower() == queryRole.lower():
-                members = role.members
-                break
-
-        if (members is None):
-            return
-
-        names = list(map(lambda m: str(m) + "\n", members))
-        header = "List of users in {role} role - {num}".format(role=role.name,
-                                                               num=len(names))
-
-        # TODO remove for paginator take empty list for embed
-        if (len(names) == 0):
-            em = discord.Embed(title=header, colour=0xDA291C)
-            em.set_footer(text="Page 01 of 01")
-            await ctx.send(embed=em)
-        else:
-            pages = Pages(ctx,
-                          item_list=names,
-                          title=header,
-                          display_option=(3, 20),
-                          editable_content=False)
-            await pages.paginate()
 
     @commands.command()
     async def version(self, ctx):
